revisit cryptfs support
With this commit, luks partition is created and mapped by systemd, not
in installer or initramfs anymore.

All partitions intended to be created as LUKS are supposed to have a
prefix label "luks_", such as

/dev/sdb3: LABEL="luks_fluxdata" UUID="6385a7ad-7831-4b6f-98ed-7758aa2dc470" TYPE="crypto_LUKS" PARTUUID="27735fec-03"

After the LUKS is created and mapped, the prefix label will be trimmed,
like this:

/dev/sdb3: UUID="6385a7ad-7831-4b6f-98ed-7758aa2dc470" TYPE="crypto_LUKS" PARTUUID="27735fec-03"
/dev/mapper/luks_sdb3: LABEL="fluxdata" UUID="937f120d-b14a-4818-84c0-896fbfe45864" TYPE="ext4"

This feature works with TPM2.0 module only for now.

If no TPM2.0 module available on the target, all partitions with luks prefix label
"luks_xxx" will be relabed to "xxx".

Signed-off-by: Yunguo Wei <yunguo.wei@windriver.com>